fix: Move isTesting to helpers
